requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==3.0.1
requests==2.31.0
SQLAlchemy==2.0.35
psycopg
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.23.4
